requirements.txt
telethon==1.28.5
aiohttp==3.8.5
fastapi==0.104.1
uvicorn==0.24.0
jinja2==3.1.2
python-multipart==0.0.6
geopy==2.3.0
python-dotenv==1.0.0
asyncio==3.4.3
sqlite3
logging